+++ server/error.go
@@ -1,4 +1,4 @@
-package utils
+package server
 
 import (
 	"fmt"
